+++ backend/.env.development
@@ -22,3 +22,4 @@ NODEMAILER_USER=
 NODEMAILER_PASS=
 NODEMAILER_FROM=
 NODEMAILER_SUBJECT=
+NODEMAILER_MESSAGECONTACT=Mensaje de contatc
