--- backend/app/utils/messagEmail.js
@@ -1,3 +1,0 @@
-const mensaje = "Mensaje de respuesta";
-
-module.exports = mensaje;
